fix error

+++ stratum/algos/blake2b.h
@@ -7,7 +7,7 @@ extern "C" {
 
 #include <stdint.h>
 
-void blake2b_hash(const char* input, char* output, uint64_t len);
+void blake2b_hash(const char* input, char* output, uint32_t len);
 
 #ifdef __cplusplus
 }
